requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
aiomysql==0.2.0
mailjet-rest==1.3.4
psutil==7.0.0
cryptography==45.0.6
python-dateutil==2.9.0.post0
requests==2.32.3